+++ 2-Authorization-I/1-call-graph/SPA/src/pages/Contacts.jsx
@@ -2,14 +2,14 @@ import { useEffect, useState } from 'react';
 
 import { ContactsData } from '../components/DataDisplay';
 import { protectedResources } from '../authConfig';
-import useFetchWithMsal from '../hooks/useFetchWithMsal';
+import useGraphWithMsal from '../hooks/useGraphWithMsal';
 
 export const Contacts = () => {
     const [graphData, setGraphData] = useState(null);
-    const request = {
+    
+    const { error, result, execute } = useGraphWithMsal({
         scopes: protectedResources.graphContacts.scopes,
-    };
-    const { error, result, execute } = useFetchWithMsal(request, protectedResources.graphContacts.endpoint);
+    }, protectedResources.graphContacts.endpoint);
 
     useEffect(() => {
         if (!!graphData) {
