+++ 2-Authorization-I/1-call-graph/SPA/src/utils/claimUtils.js
@@ -1,3 +1,6 @@
+import { AccountInfo } from '@azure/msal-browser';
+import { addClaimsToStorage } from './storageUtils';
+
 /**
  * Populate claims table with appropriate description
  * @param {Object} claims ID token claims
@@ -224,10 +227,10 @@ const changeDateFormat = (date) => {
 
 
 /**
-   * This method parses WWW-Authenticate authentication headers 
-   * @param header
-   * @return {Object} challengeMap
-   */
+ * This method parses WWW-Authenticate authentication headers 
+ * @param header
+ * @return {Object} challengeMap
+ */
 export const parseChallenges = (header) => {
     const schemeSeparator = header.indexOf(' ');
     const challenges = header.substring(schemeSeparator + 1).split(',');
@@ -239,6 +242,46 @@ export const parseChallenges = (header) => {
     });
 
     return challengeMap;
-}
+};
 
+/**
+ * This method inspects the HTTPS response from a fetch call for the "www-authenticate header"
+ * If present, it grabs the claims challenge from the header and store it in the localStorage
+ * For more information, visit: 
+ * https://docs.microsoft.com/en-us/azure/active-directory/develop/claims-challenge#claims-challenge-header-format
+ * @param {Object} response
+ * @param {string} apiEndpoint
+ * @param {AccountInfo} account
+ * @returns response
+ */
+export const handleClaimsChallenge = async (response, apiEndpoint, account) => {
+    if (response.status === 200) {
+        return response.json();
+    }
 
+    if (response.status === 401) {
+        if (response.headers.get('WWW-Authenticate')) {
+            const authenticateHeader = response.headers.get('WWW-Authenticate');
+            const claimsChallenge = parseChallenges(authenticateHeader);
+
+            /**
+             * This method stores the claim challenge to the session storage in the browser to be used when 
+             * acquiring a token. To ensure that we are fetching the correct claim from the storage, we are 
+             * using the clientId of the application and oid (user’s object id) as the key identifier of 
+             * the claim with schema cc.<clientId>.<oid>.<resource.hostname>
+             */
+            addClaimsToStorage(
+                claimsChallenge.claims,
+                `cc.${msalConfig.auth.clientId}.${account.idTokenClaims.oid}.${new URL(apiEndpoint).hostname}`
+            );
+
+            const err = new Error("A claims challenge has occurred");
+            err.name = 'ClaimsChallengeAuthError';
+            throw err;
+        }
+
+        throw new Error(`Unauthorized: ${response.status}`);
+    } else {
+        throw new Error(`Something went wrong with the request: ${response.status}`);
+    }
+};
